Reactions update (#103)
* initial work on message records database

* implemented a new logic for storing message records in the database + implemented a new layout for reactions + implemented initial logic for a new display of reactions in the code

* merge fix

* Revert "added sent at / edited at in message details"

This reverts commit 2b0068c9c0947a30e0cd7fc389d47eb6e0193e8a.

* reaction little fixes

* pr fixes

+++ app/src/main/java/com/clover/studio/exampleapp/data/models/entity/MessageRecords.kt
@@ -12,8 +12,7 @@ import kotlinx.parcelize.Parcelize
 data class MessageRecords(
     @PrimaryKey
     @ColumnInfo(name = "id")
-    // TODO id: Long
-    val id: Int,
+    val id: Long,
 
     @ColumnInfo(name = "message_id")
     val messageId: Int,
